code clean-up moved common code to utils
